requirements.txt
aiohttp==3.8.1
bs4==0.0.1
glqiwiapi==2.15
loguru==0.6.0
pycryptodome==3.15.0
Pyrogram==2.0.52
TgCrypto==1.2.3
toml==0.10.2